Add files via upload
abmu.yml is the Ansible version of the bmu.sh Bash shell script. It uses the similar logic with slightly different syntax for Python compatibility. Since the Ansible version is much easier to understand, I decided not to split the configuration into a second file -- if you want to make changes, edit the yml file directly. Extension groups are the same with the exception of including PDF as part of documents. It's extremely simple to add additional groups -- just copy and paste an existing group to a new line and edit the type and exts values.  Don't forget to do it under both tasks!  If you don't want to exclude one of the directories, just delete the whole -not -path line.  For a detailed explanation of the find command, see the bmu manual.